requirements.txt
mcp
litellm
python-dotenv
